Now Mozart compile mostly warning-free again! + Fixed most warnings about missing const with literal strings (remaining one is for virtualString to CString which is to be reviewed) + Changed the reporting flags in the DSS to sane values (removed -Weffc++ and -Winline, forced the use of most flags for GCC > 3) + Fixed most of the remaining warnings but   - usage of tmpnam -> requires some user-visible changes   - warnings in ozbison; well it's in code that's more than 12(!) years old. The good way would be to update the underlying bison. Pain ahead. + Fixed some bugs here and there (warnings are supposed to help locate them when they aren't drown in noise) +The output of make bootstrap dropped from 1.7 Mb to 189 kb!
git-svn-id: https://gforge.info.ucl.ac.be/svn/mozart@17178 ada56829-ad1f-0410-b00f-83cda6628aec

+++ platform/dss/src/msl_dct.cc
@@ -65,7 +65,7 @@ namespace _msl_internal{
       } else
 	return false; // not done
     }
-    int len = t_min(bb->availableSpace(), static_cast<int>(a_size)); // how much can and want we to marshal
+    int len = t_min(bb->availableSpace(), a_size); // how much can and want we to marshal
     bb->writeToBuffer(a_pos, len);
     a_pos += len;
     //printf("len:%d size:%d done:%s\n",len,a_size, (getMarshaled() == a_size)?"TRUE":"FALSE");
